chore: Phase 3 — remove deployed-API surface (migration cleanup)
The hosted API (MCP / A2A / download / OAuth / signing webhook) has moved
to the private UseJunior/openagreements-org-deploy repo, served behind
openagreements.org via a legal-explainer rewrite (PR
UseJunior/legal-explainer#102). The public repo's role is now purely the
library (packages/) + content. This commit removes everything that was
specific to the openagreements.org Vercel deploy.

Deletes:
- vercel.json (deploy config)
- api/** (mcp.ts, a2a.ts, download.ts, status.ts, auth/**, signing/webhook.ts,
  and 7 _helpers used only by those endpoints)
- 6 integration tests scoped to the hosted endpoints:
  api-endpoints, api-shared, api-ratelimit, api-download-tokens,
  mcp-contract, trust-signal-surfaces

Spec changes (openspec/specs/open-agreements/spec.md):
- Removed 36 scenarios that describe hosted-API behaviors now in the
  private deploy repo (OA-DST-007, 017-025, 030-031, 036-053, 063, 081-086,
  excluding 024 and 032 which are also covered by
  packages/contract-templates-mcp tests we keep).
- Removed 9 Requirement blocks that were orphaned by the scenario removals
  (Opaque Download Links, HEAD probing, Download Token Lifecycle, Durable
  Download Storage, MCP Rate Limiting on /api/mcp, Structured Lifecycle
  Logs for MCP HTTP Transport, etc.).
- Kept scenarios about public-repo trust signals (006, 008-013) and the
  system-card rendering pending bindings (068-072) — these describe
  behaviors in this repo, not in the deploy repo.

New test file (integration-tests/repo-trust-signals.test.ts):
- Carries forward the public-repo bindings from the deleted
  trust-signal-surfaces.test.ts: real tests for OA-DST-006, 008, 009, 010,
  011+012, 013 (asserting on README.md, .github/workflows/ci.yml,
  .github/codecov.yml, vitest.config.ts).
- Carries forward the skip-with-reason bindings for OA-DST-068-072.
- Drops everything that asserted on vercel.json or the deployed endpoints.

Regenerated:
- site/trust/system-card.md (reflects 36 fewer scenarios)
- integration-tests/OPENSPEC_TRACEABILITY.md

Verified:
- npm run check:spec-coverage: PASS open-agreements 229 scenarios (15
  pending with reasons — same baseline count as before).
- npm run build: clean.
- npx vitest run integration-tests/repo-trust-signals.test.ts: 6 passed,
  5 skipped (with reasons).
- Full vitest suite: 950 passed, 23 skipped, 1 failed
  (packages/signing/tests/gcloud-storage.test.ts "stores and retrieves a
  document from GCS" — pre-existing failure on origin/main, requires live
  GCP credentials, not caused by this PR).

Closes the openagreements-org-deploy migration (Phase 3 of 3).
