--- vercel.json
@@ -1,61 +1,0 @@
-{
-  "$schema": "https://openapi.vercel.sh/vercel.json",
-  "framework": null,
-  "installCommand": "npm ci --include=dev",
-  "buildCommand": "npm run build:site:vercel",
-  "outputDirectory": "_site",
-  "cleanUrls": true,
-  "trailingSlash": false,
-  "rewrites": [
-    { "source": "/.well-known/oauth-protected-resource", "destination": "/api/auth/well-known-protected-resource" },
-    { "source": "/.well-known/oauth-authorization-server", "destination": "/api/auth/well-known-oauth-as" }
-  ],
-  "headers": [
-    {
-      "source": "/.well-known/mcp-server-card",
-      "headers": [
-        { "key": "Content-Type", "value": "application/json" },
-        { "key": "Cache-Control", "value": "public, max-age=3600" },
-        { "key": "Access-Control-Allow-Origin", "value": "*" },
-        { "key": "Access-Control-Allow-Methods", "value": "GET" },
-        { "key": "Access-Control-Allow-Headers", "value": "Content-Type" }
-      ]
-    },
-    {
-      "source": "/.well-known/api-catalog",
-      "headers": [
-        { "key": "Content-Type", "value": "application/linkset+json; profile=\"https://www.rfc-editor.org/info/rfc9727\"" },
-        { "key": "Cache-Control", "value": "public, max-age=3600" },
-        { "key": "Access-Control-Allow-Origin", "value": "*" },
-        { "key": "Access-Control-Allow-Methods", "value": "GET" },
-        { "key": "Access-Control-Allow-Headers", "value": "Content-Type" }
-      ]
-    }
-  ],
-  "redirects": [
-    { "source": "/", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/templates", "destination": "https://usejunior.com/developer-tools/open-agreements/templates", "statusCode": 301 },
-    { "source": "/templates/:name", "destination": "https://usejunior.com/developer-tools/open-agreements/templates/:name", "statusCode": 301 },
-    { "source": "/docs", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/docs/:slug", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/trust", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/trust/system-card", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/trust/template-evidence", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/trust/evidence-story", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 },
-    { "source": "/trust/changelog", "destination": "https://usejunior.com/developer-tools/open-agreements", "statusCode": 301 }
-  ],
-  "functions": {
-    "api/a2a.ts": {
-      "maxDuration": 30,
-      "includeFiles": "{content/templates,content/external,content/recipes,dist}/**"
-    },
-    "api/mcp.ts": {
-      "maxDuration": 30,
-      "includeFiles": "{content/templates,content/external,content/recipes,dist,node_modules/@usejunior/docx-core/dist,packages/signing/src}/**"
-    },
-    "api/download.ts": {
-      "maxDuration": 30,
-      "includeFiles": "{content/templates,content/external,content/recipes,dist,node_modules/@usejunior/docx-core/dist}/**"
-    }
-  }
-}
